рефакторит показ ранга пользователя
