+++ src/components/profile.js
@@ -1,4 +1,23 @@
 import AbstractComponent from "./abstract-component";
+import {getHistoryMovies} from "../utils/filter";
+
+const RANK = {
+  novice: {
+    rank: `novice`,
+    minCount: 1,
+    maxCount: 10,
+  },
+  fan: {
+    rank: `fan`,
+    minCount: 11,
+    maxCount: 20,
+  },
+  movieBuff: {
+    rank: `movie buff`,
+    minCount: 21,
+    maxCount: Infinity,
+  },
+};
 
 export const createProfileTemplate = (rank) => {
   return (
@@ -10,12 +29,20 @@ export const createProfileTemplate = (rank) => {
 };
 
 export default class Profile extends AbstractComponent {
-  constructor(rank) {
+  constructor(movies) {
     super();
-    this._rank = rank;
+    this._movies = movies;
   }
 
   getTemplate() {
-    return createProfileTemplate(this._rank);
+    return createProfileTemplate(this.getUserRank());
+  }
+
+  getUserRank() {
+    const countFilms = getHistoryMovies(this._movies).length;
+    const ranks = Object.values(RANK);
+    const userRank = ranks.filter(({minCount, maxCount, rank}) => (countFilms >= minCount && countFilms <= maxCount) ? rank : ``);
+
+    return userRank;
   }
 }
